Crate prefixmap

Prefix map implementation

Implements prefix maps, which are used in Turtle and ShEx

A prefix map is a list of alias declarations associated with IRIs

prefix schema: <http://schema.org/>
prefix :       <http://example.org/>

Example

let schema_iri  = IriS::from_str("http://schema.org/")?;
let example_iri = IriS::from_str("http://example.org/")?;
let mut pm = PrefixMap::new();
pm.insert("schema", &schema_iri);
pm.insert("", &example_iri);

Re-exports

pub use crate::prefixmap::*;

pub use alias::*;

pub use deref::*;

pub use iri_ref::*;

pub use prefixmap_error::*;

Modules

alias

deref

iri_ref

prefixmap

prefixmap_error